+++ src/opt.rs
@@ -14,6 +14,8 @@ pub const PASSTHROUGH_SUBCLASS: Opt = 1 << 8;
 pub const PASSTHROUGH_DATETIME: Opt = 1 << 9;
 pub const APPEND_NEWLINE: Opt = 1 << 10;
 pub const PASSTHROUGH_DATACLASS: Opt = 1 << 11;
+pub const SERIALIZE_SET: Opt = 1 << 12;
+pub const SERIALIZE_GENERATOR: Opt = 1 << 13;
 
 // deprecated
 pub const SERIALIZE_DATACLASS: Opt = 0;
@@ -34,6 +36,8 @@ pub const MAX_OPT: i32 = (APPEND_NEWLINE
     | PASSTHROUGH_SUBCLASS
     | SERIALIZE_DATACLASS
     | SERIALIZE_NUMPY
+    | SERIALIZE_SET
+    | SERIALIZE_GENERATOR
     | SERIALIZE_UUID
     | SORT_KEYS
     | STRICT_INTEGER